+++ dbt_dry_run/cli.py
@@ -25,6 +25,7 @@ def dry_run(
     cli_vars: str = "{}",
     skip_not_compiled: bool = False,
     extra_check_columns_metadata_key: Optional[str] = None,
+    threads: Optional[int] = None,
 ) -> int:
     cli_vars_parsed = json.loads(cli_vars)
     set_flags(
@@ -38,6 +39,7 @@ def dry_run(
         profiles_dir=os.path.abspath(profiles_dir),
         target=target,
         vars=cli_vars_parsed,
+        threads=threads,
     )
     project = ProjectService(args)
     exit_code: int
@@ -71,6 +73,11 @@ _EXTRA_CHECK_COLUMNS_METADATA_KEY_HELP = """
     or it will be cast to a boolean to be 'True/Falsey`
 """
 
+_THREADS_HELP = """
+"[dbt] Number of threads to execute DAG with. You can normally set this higher than the concurrency of your actual dbt
+runs because the dry run queries execute much faster and don't use any resources"
+"""
+
 
 def version_callback(value: bool) -> None:
     if value:
@@ -88,6 +95,7 @@ def run(
     ),
     vars: str = Option("{}", help="[dbt] CLI Variables to pass to dbt"),
     target: Optional[str] = Option(None, help="[dbt] Target profile"),
+    threads: Optional[int] = Option(None, help=_THREADS_HELP),
     verbose: bool = Option(False, help="Output verbose error messages"),
     report_path: Optional[str] = Option(None, help="Json path to dump report to"),
     skip_not_compiled: bool = Option(
@@ -109,6 +117,7 @@ def run(
         vars,
         skip_not_compiled,
         extra_check_columns_metadata_key,
+        threads,
     )
     if exit_code > 0:
         raise typer.Exit(exit_code)